+++ components/AdminProfileCard.jsx
@@ -1,6 +1,7 @@
 import classNames from 'classnames'
 import {useState} from 'react'
 import Modal from 'react-modal'
+import AdminProfileForm from './AdminProfileForm'
 
 export default function KoderProfileCard({ admin }) {
     const [activeModal, setActiveModal] = useState (false)     
@@ -12,7 +13,8 @@ export default function KoderProfileCard({ admin }) {
     }
   
     return (
-      
+    <div>
+				<h3>Perfil</h3>
         <div className={classNames(
           'bg-black-ka',
           'border border-white-ka',
@@ -37,72 +39,7 @@ export default function KoderProfileCard({ admin }) {
             'text-white-ka'
           )}  
         >
-          <div className={classNames(
-            'max-w-xs',' p-10'
-            )}>
-            <form>
-              <div className={classNames('mb-4')}>
-                <label classNameName={classNames(
-                  'label-modal'
-                  )} for='firstName'>
-                    Nombre
-                </label>
-                <input className={classNames('input-modal')}
-                  id='username' type='text' placeholder='Nombre'/> 
-              </div>
-              <div className={classNames('mb-4')}>
-                <label className={classNames(
-                  'label-modal'
-                  )} for='lastName'>
-                    Apellido
-                </label>
-                <input className={classNames('input-modal')}
-                  id='username' type='text' placeholder='Apellido'/> 
-              </div>
-              <div className={classNames('mb-4')}>
-                <label className={classNames(
-                  'label-modal'
-                  )} for='imagen'>
-                    Tu imagen
-                </label>
-                <input className={classNames(
-                  'input-modal'
-                  )} id="username" type="text" placeholder="Tu foto"/> 
-              </div>
-              <div className={classNames('mb-4')}>
-                <label className={classNames(
-                  'label-modal')} for='text'>
-                  Github
-                </label>
-                <input className={classNames(
-                  'input-modal'
-                  )} id='github' type='text' placeholder='name@github.com'/>
-                <p className={classNames(
-                  'text-white-ka text-xs'
-                  )}>Ingresa tu github</p>
-              </div>
-              <div className={classNames(
-                'flex ',
-                'items-center',
-                'justify-between'
-                )}>
-                <button className={classNames(
-                  'button-ka',
-                  'hover:bg-cyan-ka'
-                  )} type="button" >
-                  Enviar
-                </button>
-                <a className={classNames(
-                  'align-baseline', 
-                  'inline-block',
-                  'hover:text-cyan-ka',
-                  'text-sm text-white-ka'
-                  )} href="#" onClick={closeModal}>
-                  Cancelar
-                </a>
-      </div>
-    </form>
-  </div>
+          <AdminProfileForm closeModal={closeModal}/>
         </Modal>
       <div className={classNames(
         'mx-auto',
@@ -139,11 +76,6 @@ export default function KoderProfileCard({ admin }) {
             'text-xs',
             ' lg:text-base'
           )}>
-            <i  className={classNames(
-              'fab fa-github',
-              ' mr-2' 
-            )}/>
-            
           </span>
           <p>{admin.location}</p>
           <p className={classNames(
@@ -151,10 +83,10 @@ export default function KoderProfileCard({ admin }) {
             'pt-0',
             'text-cyan-ka'
           )}>
-            
+            Administradora
           </p>
       </div>
   </div>
-  
+  </div>
   )
   }
